Lisansüstü/Kollektif Öğrenme/proje/2025/hüsamettin_ışıktaş/requirements.txt
torch
torchvision
open_clip_torch
timm
kaggle
scikit-learn
pandas
numpy
matplotlib
Pillow
tqdm
requests
